Function libtest_mimic::run

pub fn run(args: &Arguments, tests: Vec<Trial>) -> Conclusion

Runs all given trials (tests & benchmarks).

This is the central function of this crate. It provides the framework for the testing harness. It does all the printing and house keeping.

The returned value contains a couple of useful information. See Conclusion for more information. If --list was specified, a list is printed and a dummy Conclusion is returned.